+++ lib/features/working_directory/presentation/bloc/working_directory_state.dart
@@ -17,6 +17,8 @@ enum WorkingDirectoryBlocStatus {
   stashDropped,
   askForDiscardAllChanges,
   askForDiscardFileChanges,
+  fileChangesDiscarded,
+  allChangesDiscarded,
   addingCommits,
   amendingCommit,
   commitsAdded,